requirements.txt
openai>=1.30.0
Pillow>=10.0.0
opencv-python>=4.8.0
rembg>=2.0.50
ffmpeg-python>=0.2.0
langdetect>=1.0.9
python-dotenv>=1.0.0
pytest>=8.0.0
pyyaml>=6.0.0
onnxruntime>=1.16.0
pdf2image>=1.16.0
